\name{road.map}
\alias{road.map}

\title{Prints the road map of running kfsst for a track}
\description{
  This is a pure documentation function. Once invoked it 
  outlines the steps needed to run the SST aided tracking 
  model.      
}
\usage{
road.map()
}

\value{
  No value is returned. This function is invoked for its side effects.
}

\author{Anders Nielsen \email{anders.nielsen@hawaii.edu}, John Sibert \email{sibert@hawaii.edu}}

\seealso{\code{\link{kfsst}}, \code{\link{blue.shark}}}

\keyword{models}
