\name{ukfsst-package}
\alias{ukfsst-package}
\alias{ukfsst}
\docType{package}
\title{
Unscented Kalman filter tracking of marine animals 
}
\description{
Given a dataset of date, lon, lat, and tag-measured sea surface temperature 
this uses the unscented Kalman filter to estimate the most probable track 
of the tagged animal. 
}
\details{
\tabular{ll}{
Package: \tab ukfsst\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2006-12-11\cr
License: \tab BSD\cr
}
See the blue.shark example
}

\author{Anders Nielsen \email{anielsen@dina.kvl.dk}, 
        John Sibert \email{sibert@hawaii.edu}, and 
        Chi Lam \email{chihinl@usc.edu}}

\keyword{ package }

\seealso{\code{\link{road.map}}, \code{link{blue.shark}}}

